requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.1
cors==1.0.1
cycler==0.12.1
dj-database-url==3.0.0
Django==5.1.7
django-recaptcha
filelock==3.18.
fonttools==4.55.2
future==1.0.0
futures==3.0.5
gevent
greenlet==3.2.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
marshmallow==4.0.
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-file==2.1.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tldextract==5.3.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2